// integration/use-map-type/google/protobuf/timestamp.ts
export interface Timestamp {
/**
 * Represents seconds of UTC time since Unix epoch
 * 1970-01-01T00:00:00Z. Must be from 0001-01-01T00:00:00Z to
 * 9999-12-31T23:59:59Z inclusive.
 */
seconds: number,
/**
 * Non-negative fractions of a second at nanosecond resolution. Negative
 * second values with fractions must still have non-negative nanos values
 * that count forward in time. Must be from 0 to 999,999,999
 * inclusive.
 */
nanos: number,
}

function createBaseTimestamp(): Timestamp {
      return { seconds: 0,nanos: 0 };
    }

export const Timestamp: MessageFns<Timestamp> = {
              encode(
      message: Timestamp,
      writer: BinaryWriter = new BinaryWriter(),
    ): BinaryWriter {
if ( message.seconds !== 0) {
          writer.uint32(8).int64(message.seconds);
        }
if ( message.nanos !== 0) {
          writer.uint32(16).int32(message.nanos);
        }
return writer;
},

decode(
      input: BinaryReader | Uint8Array,
      length?: number,
    ): Timestamp {
      const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
      let end = length === undefined ? reader.len : reader.pos + length;
const message = createBaseTimestamp();
while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
case 1: {
if (tag !== 8) {
        break;
      }
    
        message.seconds = longToNumber(reader.int64());
continue; }
case 2: {
if (tag !== 16) {
        break;
      }
    
        message.nanos = reader.int32();
continue; }
}
if ((tag & 7) === 4 || tag === 0) {
        break;
      }
reader.skip(tag & 7);
}
return message;
},

fromJSON(object: any): Timestamp {
      return {
seconds: isSet(object.seconds)
          ? globalThis.Number(object.seconds)
          : 0,
nanos: isSet(object.nanos)
          ? globalThis.Number(object.nanos)
          : 0,
};
},

toJSON(message: Timestamp): unknown {
      const obj: any = {};
if ( message.seconds !== 0) {
          obj.seconds = Math.round(message.seconds);
        }
if ( message.nanos !== 0) {
          obj.nanos = Math.round(message.nanos);
        }
return obj;
},

create<I extends Exact<DeepPartial<Timestamp>, I>>(base?: I): Timestamp {
        return Timestamp.fromPartial(base ?? ({} as any));
      },
fromPartial<I extends Exact<DeepPartial<Timestamp>, I>>(object: I): Timestamp {
const message = createBaseTimestamp();
message.seconds = object.seconds ?? 0;
message.nanos = object.nanos ?? 0;
return message;
}
            };







type Builtin = Date | Function | Uint8Array | string | number | boolean |  | undefined;

export type DeepPartial<T> =  T extends Builtin
        ? T
        
        : T extends globalThis.Array<infer U>
        ? globalThis.Array<DeepPartial<U>>
        : T extends ReadonlyArray<infer U>
        ? ReadonlyArray<DeepPartial<U>>
        : T extends {}
        ? { [K in keyof T]?: DeepPartial<T[K]> }
        : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
      export type Exact<P, I extends P> = P extends Builtin
        ? P
        : P &
        { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P> >]: never };
// ...
export interface MessageFns<T> {
        encode(message: T, writer?: BinaryWriter): BinaryWriter;
decode(input: BinaryReader | Uint8Array, length?: number): T;
fromJSON(object: any): T;
toJSON(message: T): unknown;
create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
      }